backend/functions/requirements.txt
firebase_functions~=0.1.0
firebase-admin>=6.5.0
google-cloud-firestore>=2.0.0
google-genai>=0.2.0
python-dotenv>=1.0.0